Adding support for nested nav stacks.
One of the tests also discovered that we haven't implemented rendering of the page in the target DOM element, so this also happens in this commit. To make the tests not fail, we need to spy instead of stub Navstack.renderPage, so that we don't attempt to appendChild with a null dom element.