Only the last page in the stack is checked for being abstract.
In effect, this means that only the page we actually navigate too, i.e. by pushPage or the last page in a .navigate, is the only one we check for being abstract and give special treatment. Any other abstract page will get .route called on it via the .navigate anyway.